requirements.txt
pandas
fuzzywuzzy
python-Levenshtein
mysql-connector-python
flask